Data-layer: fix setting apiNamespace in requests
If passed to `http()`, the `apiNamespace` parameter was being included
in the `query` parameter passed to `wpcom.req.post()`, but not in the
`params` parameter where it needs to be to affect the request path.

This change always includes `apiNamespace` in the `params`, which will
allow the path to be correctly changed.

It's likely that there's another way to do this, but I'm not familiar
with the `query` param of `wpcom.req.post()` and it seems to be
undocumented in the wpcom.js source.